Move email list functionality out into a custom hook
